Client/requirements.txt
hashlib
os
pycryptodome
sys
requests
rsa